✅(tooling-domain) move entity in ddd and add test for max coverage
