+++ src/web/domain/ddd/entity.py
